+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Geometry primitives for composite layout: ranges, grids, alignment, margins
+- Vectorized linear/log/lin-log coordinate transforms and line/text/theme style objects
 
 ### Changed